SEQNG-256 Better organize the client modules
