+++ modules/edu.gemini.seqexec.web/edu.gemini.seqexec.web.client/src/main/scala/edu/gemini/seqexec/web/client/SeqexecApp.scala
@@ -5,7 +5,9 @@ package edu.gemini.seqexec.web.client
 
 import edu.gemini.seqexec.web.client.components.{SeqexecStyles, SeqexecUI}
 import edu.gemini.seqexec.web.client.services.log.{AjaxHandler, ConsoleHandler}
-import edu.gemini.seqexec.web.client.model.{Initialize, Pages, SeqexecCircuit}
+import edu.gemini.seqexec.web.client.model.Pages
+import edu.gemini.seqexec.web.client.actions.Initialize
+import edu.gemini.seqexec.web.client.circuit.SeqexecCircuit
 import edu.gemini.seqexec.model.Model.SeqexecSite
 import org.scalajs.dom.document
