+++ modules/edu.gemini.seqexec.web/edu.gemini.seqexec.web.client/src/main/scala/edu/gemini/seqexec/web/client/components/TopMenu.scala
@@ -4,7 +4,8 @@
 package edu.gemini.seqexec.web.client.components
 
 import diode.react.ModelProxy
-import edu.gemini.seqexec.web.client.model.{ClientStatus, Logout, OpenLoginBox}
+import edu.gemini.seqexec.web.client.actions.{Logout, OpenLoginBox}
+import edu.gemini.seqexec.web.client.circuit.ClientStatus
 import edu.gemini.seqexec.web.client.semanticui.Size
 import edu.gemini.seqexec.web.client.semanticui.elements.button.Button
 import edu.gemini.seqexec.web.client.semanticui.elements.icon.Icon.IconSignOut
