+++ modules/edu.gemini.seqexec.web/edu.gemini.seqexec.web.client/src/main/scala/edu/gemini/seqexec/web/client/model/model.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2016-2017 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package edu.gemini.seqexec.web.client.model
+package edu.gemini.seqexec.web.client
 
 import diode.RootModelR
 import diode.data.{Empty, Pot, RefTo}
@@ -11,127 +11,133 @@ import edu.gemini.seqexec.model.Model.SeqexecEvent.ServerLogMessage
 import edu.gemini.web.common.FixedLengthBuffer
 import org.scalajs.dom.WebSocket
 
+import monocle.macros.Lenses
+
 import scalaz._
 import Scalaz._
 
-// Pages
-object Pages {
-  sealed trait SeqexecPages
+@SuppressWarnings(Array("org.wartremover.warts.PublicInference"))
+object model {
 
-  case object Root extends SeqexecPages
-  final case class InstrumentPage(instrument: Instrument, obsId: Option[SequenceId]) extends SeqexecPages
+  // Pages
+  object Pages {
+    sealed trait SeqexecPages
 
-  implicit val equal: Equal[SeqexecPages] = Equal.equalA
-}
+    case object Root extends SeqexecPages
+    final case class InstrumentPage(instrument: Instrument, obsId: Option[SequenceId]) extends SeqexecPages
 
-// UI model
-sealed trait SectionVisibilityState
-case object SectionOpen extends SectionVisibilityState
-case object SectionClosed extends SectionVisibilityState
-
-object SectionVisibilityState {
-  implicit val eq: Equal[SectionVisibilityState] = Equal.equalA[SectionVisibilityState]
-}
-
-final case class SequenceTab(instrument: Instrument, currentSequence: RefTo[Option[SequenceView]], completedSequence: Option[SequenceView], stepConfigDisplayed: Option[Int]) {
-  // Returns the current sequence or if empty the last completed one
-  // This must be a def since it will do a call to dereference a RefTo
-  def sequence: Option[SequenceView] = currentSequence().orElse(completedSequence)
-}
-
-object SequenceTab {
-  val empty: SequenceTab = SequenceTab(Instrument.F2, RefTo(new RootModelR(None)), None, None)
-}
-
-// Model for the tabbed area of sequences
-final case class SequencesOnDisplay(instrumentSequences: Zipper[SequenceTab]) {
-  def withSite(site: SeqexecSite): SequencesOnDisplay =
-    SequencesOnDisplay(site.instruments.map(SequenceTab(_, SequencesOnDisplay.emptySeqRef, None, None)).toZipper)
-
-  // Display a given step on the focused sequence
-  def showStep(i: Int): SequencesOnDisplay =
-    copy(instrumentSequences = instrumentSequences.modify(_.copy(stepConfigDisplayed = Some(i))))
-
-  // Don't show steps for the sequence
-  def unshowStep: SequencesOnDisplay =
-    copy(instrumentSequences = instrumentSequences.modify(_.copy(stepConfigDisplayed = None)))
-
-  def focusOnSequence(s: RefTo[Option[SequenceView]]): SequencesOnDisplay = {
-    // Replace the sequence for the instrument or the completed sequence
-    val q = instrumentSequences.findZ(i => s().exists(_.metadata.instrument === i.instrument)).map(_.modify(_.copy(currentSequence = s)))
-    copy(instrumentSequences = q | instrumentSequences)
+    implicit val equal: Equal[SeqexecPages] = Equal.equalA
   }
 
-  def focusOnInstrument(i: Instrument): SequencesOnDisplay = {
-    // Focus on the instrument
-    val q = instrumentSequences.findZ(s => s.instrument === i)
-    copy(instrumentSequences = q | instrumentSequences)
+  // UI model
+  sealed trait SectionVisibilityState
+  case object SectionOpen extends SectionVisibilityState
+  case object SectionClosed extends SectionVisibilityState
+
+  object SectionVisibilityState {
+    implicit val eq: Equal[SectionVisibilityState] = Equal.equalA[SectionVisibilityState]
   }
 
-  def isAnySelected: Boolean = instrumentSequences.toStream.exists(_.sequence.isDefined)
-
-  // Is the id on the sequences area?
-  def idDisplayed(id: SequenceId): Boolean =
-    instrumentSequences.withFocus.toStream.find { case (s, a) => a && s.sequence.exists(_.id === id)}.isDefined
-
-  def instrument(i: Instrument): (SequenceTab, Boolean) =
-    // The getOrElse shouldn't be called as we have an element per instrument
-    instrumentSequences.withFocus.toStream.find(_._1.instrument === i).getOrElse((SequenceTab.empty, false))
-
-  // We'll set the passed SequenceView as completed for the given instruments
-  def markCompleted(completed: SequenceView): SequencesOnDisplay = {
-    val q = instrumentSequences.findZ(s => s.instrument === completed.metadata.instrument).map(_.modify(_.copy(completedSequence = completed.some)))
-    copy(instrumentSequences = q | instrumentSequences)
+  final case class SequenceTab(instrument: Instrument, currentSequence: RefTo[Option[SequenceView]], completedSequence: Option[SequenceView], stepConfigDisplayed: Option[Int]) {
+    // Returns the current sequence or if empty the last completed one
+    // This must be a def since it will do a call to dereference a RefTo
+    def sequence: Option[SequenceView] = currentSequence().orElse(completedSequence)
   }
-}
 
-/**
-  * Contains the sequences displayed on the instrument tabs. Note that they are references to sequences on the Queue
-  */
-object SequencesOnDisplay {
-  val emptySeqRef: RefTo[Option[SequenceView]] = RefTo(new RootModelR(None))
+  object SequenceTab {
+    val empty: SequenceTab = SequenceTab(Instrument.F2, RefTo(new RootModelR(None)), None, None)
+  }
 
-  // We need to initialize the model with some instruments but it will be shortly replaced by the actual list
-  val empty: SequencesOnDisplay = SequencesOnDisplay(Instrument.gsInstruments.map(SequenceTab(_, emptySeqRef, None, None)).toZipper)
-}
+  // Model for the tabbed area of sequences
+  final case class SequencesOnDisplay(instrumentSequences: Zipper[SequenceTab]) {
+    def withSite(site: SeqexecSite): SequencesOnDisplay =
+      SequencesOnDisplay(site.instruments.map(SequenceTab(_, SequencesOnDisplay.emptySeqRef, None, None)).toZipper)
 
-final case class WebSocketConnection(ws: Pot[WebSocket], nextAttempt: Int)
+    // Display a given step on the focused sequence
+    def showStep(i: Int): SequencesOnDisplay =
+      copy(instrumentSequences = instrumentSequences.modify(_.copy(stepConfigDisplayed = Some(i))))
 
-object WebSocketConnection {
-  val empty: WebSocketConnection = WebSocketConnection(Empty, 0)
-}
+    // Don't show steps for the sequence
+    def unshowStep: SequencesOnDisplay =
+      copy(instrumentSequences = instrumentSequences.modify(_.copy(stepConfigDisplayed = None)))
 
-/**
-  * Keeps a list of log entries for display
-  */
-final case class GlobalLog(log: FixedLengthBuffer[ServerLogMessage])
+    def focusOnSequence(s: RefTo[Option[SequenceView]]): SequencesOnDisplay = {
+      // Replace the sequence for the instrument or the completed sequence
+      val q = instrumentSequences.findZ(i => s().exists(_.metadata.instrument === i.instrument)).map(_.modify(_.copy(currentSequence = s)))
+      copy(instrumentSequences = q | instrumentSequences)
+    }
 
-/**
- * UI model, changes here will update the UI
- */
-final case class SeqexecUIModel(navLocation: Pages.SeqexecPages,
-                          user: Option[UserDetails],
-                          sequences: SeqexecAppRootModel.LoadedSequences,
-                          loginBox: SectionVisibilityState,
-                          globalLog: GlobalLog,
-                          sequencesOnDisplay: SequencesOnDisplay,
-                          firstLoad: Boolean)
+    def focusOnInstrument(i: Instrument): SequencesOnDisplay = {
+      // Focus on the instrument
+      val q = instrumentSequences.findZ(s => s.instrument === i)
+      copy(instrumentSequences = q | instrumentSequences)
+    }
 
-object SeqexecUIModel {
-  val noSequencesLoaded: SequencesQueue[SequenceView] = SequencesQueue[SequenceView](Conditions.default, None, Nil)
-  val initial: SeqexecUIModel = SeqexecUIModel(Pages.Root, None, noSequencesLoaded,
-    SectionClosed, GlobalLog(FixedLengthBuffer.unsafeFromInt(100)), SequencesOnDisplay.empty, true)
-}
+    def isAnySelected: Boolean = instrumentSequences.toStream.exists(_.sequence.isDefined)
+
+    // Is the id on the sequences area?
+    def idDisplayed(id: SequenceId): Boolean =
+      instrumentSequences.withFocus.toStream.find { case (s, a) => a && s.sequence.exists(_.id === id)}.isDefined
+
+    def instrument(i: Instrument): (SequenceTab, Boolean) =
+      // The getOrElse shouldn't be called as we have an element per instrument
+      instrumentSequences.withFocus.toStream.find(_._1.instrument === i).getOrElse((SequenceTab.empty, false))
+
+    // We'll set the passed SequenceView as completed for the given instruments
+    def markCompleted(completed: SequenceView): SequencesOnDisplay = {
+      val q = instrumentSequences.findZ(s => s.instrument === completed.metadata.instrument).map(_.modify(_.copy(completedSequence = completed.some)))
+      copy(instrumentSequences = q | instrumentSequences)
+    }
+  }
+
+  /**
+    * Contains the sequences displayed on the instrument tabs. Note that they are references to sequences on the Queue
+    */
+  object SequencesOnDisplay {
+    val emptySeqRef: RefTo[Option[SequenceView]] = RefTo(new RootModelR(None))
+
+    // We need to initialize the model with some instruments but it will be shortly replaced by the actual list
+    val empty: SequencesOnDisplay = SequencesOnDisplay(Instrument.gsInstruments.map(SequenceTab(_, emptySeqRef, None, None)).toZipper)
+  }
+
+  @Lenses final case class WebSocketConnection(ws: Pot[WebSocket], nextAttempt: Int)
+
+  object WebSocketConnection {
+    val empty: WebSocketConnection = WebSocketConnection(Empty, 0)
+  }
+
+  /**
+    * Keeps a list of log entries for display
+    */
+  final case class GlobalLog(log: FixedLengthBuffer[ServerLogMessage])
+
+  /**
+   * UI model, changes here will update the UI
+   */
+  final case class SeqexecUIModel(navLocation: Pages.SeqexecPages,
+                            user: Option[UserDetails],
+                            sequences: SeqexecAppRootModel.LoadedSequences,
+                            loginBox: SectionVisibilityState,
+                            globalLog: GlobalLog,
+                            sequencesOnDisplay: SequencesOnDisplay,
+                            firstLoad: Boolean)
+
+  object SeqexecUIModel {
+    val noSequencesLoaded: SequencesQueue[SequenceView] = SequencesQueue[SequenceView](Conditions.default, None, Nil)
+    val initial: SeqexecUIModel = SeqexecUIModel(Pages.Root, None, noSequencesLoaded,
+      SectionClosed, GlobalLog(FixedLengthBuffer.unsafeFromInt(100)), SequencesOnDisplay.empty, true)
+  }
 
 
 
-/**
-  * Root of the UI Model of the application
-  */
-final case class SeqexecAppRootModel(ws: WebSocketConnection, site: SeqexecSite, uiModel: SeqexecUIModel)
+  /**
+    * Root of the UI Model of the application
+    */
+  final case class SeqexecAppRootModel(ws: WebSocketConnection, site: SeqexecSite, uiModel: SeqexecUIModel)
 
-object SeqexecAppRootModel {
-  type LoadedSequences = SequencesQueue[SequenceView]
+  object SeqexecAppRootModel {
+    type LoadedSequences = SequencesQueue[SequenceView]
 
-  val initial: SeqexecAppRootModel = SeqexecAppRootModel(WebSocketConnection.empty, SeqexecSite.SeqexecGS, SeqexecUIModel.initial)
+    val initial: SeqexecAppRootModel = SeqexecAppRootModel(WebSocketConnection.empty, SeqexecSite.SeqexecGS, SeqexecUIModel.initial)
+  }
 }
